+++ src/tui/view/component/help.rs
@@ -42,7 +42,7 @@ impl Draw for HelpFooter {
         frame.render_widget(
             Paragraph::new(text)
                 .alignment(Alignment::Right)
-                .style(tui_context.theme.text_highlight),
+                .style(tui_context.theme.text.highlight),
             area,
         );
     }